+++ DelvUI/changelog.md
@@ -7,6 +7,9 @@ Features:
     + A new "Performance" mode was added which has the clipping functionallity reduced in favor of FPS.
     + Details on all the modes can be found in Misc > Window Clipping.
 
+ Fixes:
+ - Fix "Change Alpha Based on Range" for Missing Health Color.
+
 # 0.6.3.4
 - Fixed window clipping not working.