+++ DelvUI/changelog.md
@@ -8,6 +8,7 @@ Features:
     + Details on all the modes can be found in Misc > Window Clipping.
 
 Fixes:
+- Fixed "Change Alpha Based on Range" for Missing Health Color.
 - Updated the pre-populated white list for the Custom Effects. Note that this will not update existing profiles.
 
 # 0.6.3.4